requirements.txt
chromadb>=0.5.5
PyMuPDF>=1.24.10
sentence-transformers>=3.0.1
streamlit>=1.37.0
pydantic>=2.8.2
python-dateutil>=2.9.0.post0
requests>=2.32.3
python-dotenv>=1.0.1
